it's cool but not all objects seem to be properly linked, we'll see

+++ libs/ft_putchar.c
@@ -5,14 +5,14 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: vcedraz- <vcedraz-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/22 13:44:15 by vcedraz-          #+#    #+#             */
-/*   Updated: 2022/11/28 20:50:29 by vcedraz-         ###   ########.fr       */
+/*   Created: 2022/11/15 14:19:54 by vcedraz-          #+#    #+#             */
+/*   Updated: 2022/11/18 18:00:41 by vcedraz-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-long	ft_putchar(char c)
+size_t	ft_putchar(char c)
 {
 	return (write(1, &c, 1));
 }
